requirements.txt
PySide6>=6.6.0
pyodbc>=4.0.0
reportlab>=4.0.0
python-barcode>=0.15.0
pillow>=10.0.0
